% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathwayAnnotation-class.R
\name{createPathwayAnnotationObject}
\alias{createPathwayAnnotationObject}
\title{Create a pathwayAnnotation object}
\usage{
createPathwayAnnotationObject(
  database,
  annotation,
  definition,
  organism = NULL,
  ensembl.version = NULL,
  annotation.version = NULL
)
}
\arguments{
\item{database}{Name of the annotation source.}

\item{annotation}{Named list where each list item is named by gene ID and contains vector of pathway IDs.}

\item{definition}{Data frame containing the columns \code{pathway_id}, \code{pathway_name}, and optionally additional columns.}

\item{organism}{Name of the organism. It is advisable to use ENSEMBL codes, which can be retrieved using \code{fetchOrganisms('ENSEMBL')}.}

\item{ensembl.version}{ENSEMBL version}

\item{annotation.version}{Annotation version}
}
\value{
An object of pathwayAnnotation class
}
\description{
This function binds several data structures together into a single convenient object,
including an annotation object (a named list named by gene ID and containing a vector
of associated pathway IDs) and a definitions object (a data frame containing the
columns "pathway_id" and "pathway_name".
}
