% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathwayAnnotation-tools.R
\name{exportGMT}
\alias{exportGMT}
\title{Export an annotation into GMT (gene matrix transposed) file format.}
\usage{
exportGMT(
  annotation,
  description = "pathway_name",
  file = "annotation_export.gmt",
  n.cores = NULL
)
}
\arguments{
\item{annotation}{An object of pathwayAnnotation class.}

\item{description}{If character class, column (of definition object) from which to populate the optional description file. If \code{NULL}, the column will be populated with dummy ("na") values.}

\item{file}{File path for exported GMT file.}

\item{n.cores}{A number specifying the number of cores to use.}
}
\description{
This function provides a workflow for exported an annotation as a GMT file. Exported files
may be used with tools such as g:Profiler.
}
