% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathwayAnnotation-tools.R
\name{fetchAnnotations}
\alias{fetchAnnotations}
\title{Retrieve current versions of annotations from online sources.}
\usage{
fetchAnnotations(
  database = c("GO", "KEGG", "PANTHER", "DISEASES", "DisGeNET", "IID"),
  ensembl.species = "hsapiens",
  kegg.species = "hsa",
  panther.species = "HUMAN",
  iid.species = "human",
  go.namespace = NULL,
  diseases.confidence = NULL,
  disgenet.class = NULL,
  disgenet.source = NULL,
  disgenet.score = NULL,
  iid.n.methods = NULL,
  iid.n.publications = NULL,
  iid.evidence.type = NULL,
  iid.direction = NULL,
  ensembl.version = NULL,
  n.cores = NULL
)
}
\arguments{
\item{database}{A character string specifying the name of the annotation to retrieve. If multiple values are given, only the first is used. Valid values are \code{"GO"}, \code{"KEGG"}, \code{"PANTHER"}, \code{"DISEASES"}, \code{"DisGeNET"}, and \code{"IID"}.}

\item{ensembl.species}{A character string specifying the organism for which genes (ENSEMBL) should be annotated. Valid values can be found using \code{fetchOrganisms("ENSEMBL")}.}

\item{kegg.species}{A character string specifying the organism from which to derive KEGG (Kyoto Encyclopedia of Genes and Genomes) annotations. Valid values can be found using \code{fetchOrganisms("KEGG")}. Used only when \code{database} is \code{"KEGG"}, ignored otherwise.}

\item{panther.species}{A character string specifying the organism from which to derive PANTHER (Protein ANalysis THrough Evolutionary Relationships) annotations. Valid values can be found using \code{fetchOrganisms("PANTHER")}. Used only when \code{database} is \code{"PANTHER"}, ignored otherwise.}

\item{iid.species}{A character string specifying the organism from which to derive IID (Integrated Interactions Database) annotations. Valid values can be found using \code{fetchOrganisms("IID")}. Used only when \code{database} is \code{"IID"}, ignored otherwise.}

\item{go.namespace}{A character string specifying the namespace for which to limit GO (Gene Ontology) annotations. Must be one or more of \code{"BP"}, \code{"CC"}, or \code{"MF"}. Used only when \code{database} is \code{"GO"}, ignored otherwise.}

\item{diseases.confidence}{A number specifying the confidence score threshold (DISEASES database) for gene-disease associations to include. If \code{NULL}, no filter is applied. Used only when \code{database} is \code{"DISEASES"}, ignored otherwise.}

\item{disgenet.class}{A character vector specifying one or more disease class(es) (MeSH codes, DisGeNET database) to include. If \code{"all"} or \code{NULL}, all classes are included. Used only when \code{database} is \code{"DisGeNET"}, ignored otherwise.}

\item{disgenet.source}{A character vector specifying one or more gene-disease association source(s) (DisGeNET database) to include. If \code{"all"} or \code{NULL}, all sources are included. Used only when \code{database} is \code{"DisGeNET"}, ignored otherwise.}

\item{disgenet.score}{A number specifying the score threshold (DisGeNET database) for gene-disease associations to include. If \code{NULL}, no filter is applied. Used only when \code{database} is \code{"DisGeNET"}, ignored otherwise.}

\item{iid.n.methods}{A number specifying the number threshold (IID database) of methods supporting a protein-protein interaction. If \code{NULL}, no filter is applied. Used only when \code{database} is \code{"IID"}, ignored otherwise.}

\item{iid.n.publications}{A number specifying the number threshold (IID database) of publications supporting a protein-protein interaction. If \code{NULL}, no filter is applied. Used only when \code{database} is \code{"IID"}, ignored otherwise.}

\item{iid.evidence.type}{A character vector specifying one or more type(s) of evidence (IID database) to include. Possible values are \code{"all"}, \code{"exp"}, \code{"pred"}, and \code{"ortho"}. If \code{"all"} or \code{NULL}, all evidence types are included. Used only when \code{database} is \code{"IID"}, ignored otherwise.}

\item{iid.direction}{A character vector specifying the direction(s) of evidence (IID database) to include. Possible values are \code{"all"}, \code{"-"}, \code{">"}, \code{"<"}, and \code{"><"}. If \code{"all"} or \code{NULL}, all directions are included. Used only when \code{database} is \code{"IID"}, ignored otherwise.}

\item{ensembl.version}{A character string specifying the version of the ENSEMBL database from which to derive annotations.}

\item{n.cores}{A number specifying the number of cores to use.}
}
\value{
An object of pathwayAnnotation class
}
\description{
This function provides specialized workflows for downloading annotations from
established databases and datasets. An internet connection is required.
}
