% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathwayAnnotation-tools.R
\name{fetchOrganisms}
\alias{fetchOrganisms}
\title{Look up valid organism codes}
\usage{
fetchOrganisms(
  database = c("ENSEMBL", "KEGG", "PANTHER", "DISEASES", "DisGeNET", "IID")
)
}
\arguments{
\item{database}{Name of the annotation source.}
}
\value{
A data frame containing an \code{org_code} column and additional metadata.
}
\description{
This function looks up valid organism values for use with specialized databases.
The valid codes are found in the \code{org_code} column of the returned data frame.
}
