% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathwayAnnotation-class.R
\name{getAnnotationsList}
\alias{getAnnotationsList}
\title{Get annotations list from an annotation object.}
\usage{
getAnnotationsList(annotation)
}
\arguments{
\item{annotation}{An object of pathwayAnnotation class.}
}
\value{
A named list with gene-to-pathway associations
}
\description{
This function extracts the annotations list from a pathwayAnnotation object.
}
