% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathwayAnnotation-class.R
\name{getAnnotationsVersion}
\alias{getAnnotationsVersion}
\title{Get pathway annotations version from an annotation object.}
\usage{
getAnnotationsVersion(annotation)
}
\arguments{
\item{annotation}{An object of pathwayAnnotation class.}
}
\value{
A string with the pathway annotations version
}
\description{
This function extracts the pathway annotations version from a pathwayAnnotation object.
}
