% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathwayAnnotation-class.R
\name{getDefinitions}
\alias{getDefinitions}
\title{Get definitions from an annotation object.}
\usage{
getDefinitions(annotation)
}
\arguments{
\item{annotation}{An object of pathwayAnnotation class.}
}
\value{
A data frame with pathway metadata
}
\description{
This function extracts the definitions data frame from a pathwayAnnotation object.
}
