% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathwayAnnotation-class.R
\name{getGenesVersion}
\alias{getGenesVersion}
\title{Get genes version from an annotation object.}
\usage{
getGenesVersion(annotation)
}
\arguments{
\item{annotation}{An object of pathwayAnnotation class.}
}
\value{
A string with the ENSEMBL genes version
}
\description{
This function extracts the ENSEMBL genes version from a pathwayAnnotation object.
}
