% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathwayAnnotation-tools.R
\name{importGMT}
\alias{importGMT}
\title{Import an annotation from GMT (gene matrix transposed) file format.}
\usage{
importGMT(file, database, organism, n.cores = NULL, ...)
}
\arguments{
\item{file}{File path for GMT file.}

\item{database}{Name of the annotation source.}

\item{organism}{Name of the organism. It is advisable to use ENSEMBL codes, which can be retrieved using \code{fetchOrganisms('ENSEMBL')}.}

\item{n.cores}{A number specifying the number of cores to use.}

\item{...}{Additional arguments to be passed to \code{createPathwayAnnotationObject}.}
}
\description{
This function provides a workflow for importing an annotation as a GMT file.
}
