% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathwayAnnotation-tools.R
\name{liftAnnotations}
\alias{liftAnnotations}
\title{Lift an annotation over to another organism on the basis of orthology information.}
\usage{
liftAnnotations(
  annotation,
  ensembl.species,
  ensembl.version = NULL,
  orthology.type = NULL,
  percent.identity.1 = NULL,
  percent.identity.2 = NULL,
  goc.score = NULL,
  wga.coverage = NULL,
  orthology.confidence = NULL,
  n.cores = NULL
)
}
\arguments{
\item{annotation}{An object of pathwayAnnotation class (query species).}

\item{ensembl.species}{A character string specifying the organism for which genes (ENSEMBL) should be annotated (target species). Valid values can be found using \code{fetchOrganisms("ENSEMBL")}.}

\item{ensembl.version}{A character string specifying the version of the ENSEMBL database from which to derive annotations for the target species.}

\item{orthology.type}{A character string specifying the type(s) of orthology allowed for transferring annotations to the target species. Possible values are \code{"ortholog_one2one"}, \code{"ortholog_one2many"}, \code{"ortholog_many2many"}. If \code{NULL} (default), \code{"ortholog_one2one"} (only single-copy orthologs) is assumed.}

\item{percent.identity.1}{A number specifying the percent identity threshold (target to query) for orthologs to include. Valid values are from \code{0} to \code{100}. If \code{NULL}, no filter is applied.}

\item{percent.identity.2}{A number specifying the percent identity threshold (query to target) for orthologs to include. Valid values are from \code{0} to \code{100}. If \code{NULL}, no filter is applied.}

\item{goc.score}{A number specifying the gene order conservation score threshold for orthologs to include. Valid values are from \code{0} to \code{100}. If \code{NULL}, no filter is applied.}

\item{wga.coverage}{A number specifying the whole genome alignment score threshold for orthologs to include. Valid values are from \code{0} to \code{100}. If \code{NULL}, no filter is applied.}

\item{orthology.confidence}{A number specifying the orthology confidence threshold for orthologs to include. Valid values are from \code{0} to \code{1}. If \code{NULL}, no filter is applied.}

\item{n.cores}{A number specifying the number of cores to use.}
}
\value{
An object of pathwayAnnotation class
}
\description{
This function provides a workflow for converting an annotation for use with a new
organism. An internet connection is required
}
