% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple-enrichment.R
\name{primEnrich}
\alias{primEnrich}
\title{Simple (\emph{prim}itive and \emph{prim}ate-inspired) enrichment analysis workflow}
\usage{
primEnrich(
  test.values,
  annotation,
  alternative = "greater",
  background = NULL,
  statistic = "fisher",
  node.size = 10,
  definition = NULL,
  n.cores = NULL
)
}
\arguments{
\item{test.values}{A vector of test values. May be a character, logical, integer, or numeric class. If a character vector is given, the \code{background} argument is required.}

\item{annotation}{Either an object of "pathwayAnnotation" class or a named list where names are gene IDs and contents are pathway IDs. If a list is given, the \code{definition} argument is required.}

\item{alternative}{A character string specifying the alternative hypothesis, must be one of \code{"greater"} (default), \code{"less"}, or \code{"two.sided"}.}

\item{background}{A character vector of background genes. Used only when \code{test.values} is a character vector, ignored otherwise.}

\item{statistic}{Character string specifing which test to use. Possible values are \code{"fisher"}, \code{"binom"}, \code{"chisq"}, \code{"ks"}, \code{"wilcox"}, and \code{"t"},}

\item{node.size}{A number representing the minimum number of genes per pathway. The default is \code{10}.}

\item{definition}{A data frame with the columns \code{pathway_id} and \code{pathway_name}, along with additional optional columns. Ignored if \code{annotation} is of the pathwayAnnotation class.}

\item{n.cores}{A number specifying the number of cores to use.}
}
\value{
A data frame of pathway IDs, test statistics, and associated pathway metadata.
}
\description{
This function provides a workflow for running simple enrichment tests using
annotations and gene lists or gene scores.
}
