% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topGO-enrichment.R
\name{topgoFisher}
\alias{topgoFisher}
\title{Wrapper for Fisher's exact test implemented in topGO}
\usage{
topgoFisher(
  test.values,
  annotation,
  background = NULL,
  node.size = 10,
  definition = NULL,
  algorithm = "weight01",
  go.ontology = "BP"
)
}
\arguments{
\item{test.values}{A vector of test values. May be a character, logical, integer, or numeric class. If a character vector is given, the \code{background} argument is required.}

\item{annotation}{Either an object of "pathwayAnnotation" class or a named list where names are gene IDs and contents are pathway IDs. If a list is given, the \code{definition} argument is required.}

\item{background}{A character vector of background genes. Used only when \code{test.values} is a character vector, ignored otherwise.}

\item{node.size}{A number representing the minimum number of genes per pathway. The default is \code{10}.}

\item{definition}{A data frame with the columns \code{pathway_id} and \code{pathway_name}, along with additional optional columns. Ignored if \code{annotation} is of the pathwayAnnotation class.}

\item{algorithm}{Character string specifing which algorithm to use. Possible values are given by \code{topGO::whichAlgorithms()}.}

\item{go.ontology}{A character string specifying the namespace for which to limit GO annotations. Must be one or more of \code{"BP"}, \code{"CC"}, or \code{"MF"}.}
}
\value{
A data frame of GO pathway IDs, test statistics, and associated metadata.
}
\description{
This function provides a workflow for running topGO enrichment tests which
take into account the topology of the GO graph.
}
